feat: add RTK api
