+++ src/services/api.ts
@@ -1,55 +1,43 @@
-import axios from 'axios';
+import { createApi, fetchBaseQuery } from '@reduxjs/toolkit/query/react';
 import { Product } from '../models/product';
 
-export const API_URL = 'https://dummyjson.com/products';
+export const BASE_URL = 'https://dummyjson.com/products';
 
 export const PRODUCTS_PER_PAGE = 15;
-
-export const SEARCH_PARAMETERS = {
-  page: 'page',
-  query: 'query',
-} as const;
 
 export interface ProductsApiResponse {
   total: number;
   products: Product[];
 }
 
-interface RequestParams {
-  query?: string;
-  page?: number;
+export interface AppSearchParams {
+  page: number;
+  query: string;
 }
 
-function createRequestUrl({ query, page }: RequestParams) {
-  const searchParams = new URLSearchParams();
+export const productApi = createApi({
+  reducerPath: 'productApi',
+  baseQuery: fetchBaseQuery({ baseUrl: BASE_URL }),
+  endpoints: (builder) => ({
+    getSearchProductsOnPage: builder.query<
+      ProductsApiResponse,
+      AppSearchParams
+    >({
+      query: (appSearchParams) => ({
+        url: `/search`,
+        params: {
+          q: appSearchParams.query,
+          limit: PRODUCTS_PER_PAGE,
+          skip: (appSearchParams.page - 1) * PRODUCTS_PER_PAGE,
+        },
+      }),
+    }),
 
-  if (page && page > 1)
-    searchParams.set('skip', ((page - 1) * PRODUCTS_PER_PAGE).toString());
-  if (query) searchParams.set('q', query);
-  searchParams.set('limit', PRODUCTS_PER_PAGE.toString());
+    getProductById: builder.query<Product, string>({
+      query: (id) => ({ url: `/${id}` }),
+    }),
+  }),
+});
 
-  return `${API_URL}${query ? '/search?' : '?'}${searchParams.toString()}`;
-}
-
-export async function getProducts(
-  requestParams: RequestParams
-): Promise<ProductsApiResponse> {
-  const requestUrl = createRequestUrl(requestParams);
-  try {
-    const response = await axios.get(requestUrl);
-    return response.data;
-  } catch (error) {
-    console.error('Error fetching data:', error);
-    throw error;
-  }
-}
-
-export async function getProductByIdPromise(id: string): Promise<Product> {
-  try {
-    const res = await fetch(`${API_URL}/${id}`);
-    return await res.json();
-  } catch (error) {
-    console.error('Error fetching data:', error);
-    throw error;
-  }
-}
+export const { useGetProductByIdQuery, useGetSearchProductsOnPageQuery } =
+  productApi;
